requirements.txt
aiogram==3.8.0
Flask==3.0.3
uvicorn==0.30.6
python-dotenv==1.0.1
SQLAlchemy[asyncio]==2.0.36
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.2
APScheduler==3.10.4
Pillow==10.4.0
ImageHash==4.3.1
pydantic==2.8.2
pydantic-settings==2.4.0
tenacity==8.5.0
loguru==0.7.2
